Migrated SqsFifoQueue receiver.
