--- .tmp/clare/clare/infrastructure/receivers.py
@@ -1,68 +1,0 @@
-# -*- coding: utf-8 -*-
-
-import collections
-
-
-class SqsFifoQueue(consumer.receivers.Receiver):
-
-    def __init__(self,
-                 sqs_queue,
-                 batch_size_maximum_count,
-                 wait_time_seconds,
-                 message_factory,
-                 _buffer=None):
-
-        """
-        Parameters
-        ----------
-        sqs_queue : boto3.resources.factory.sqs.Queue
-        batch_size_maximum_count : int
-            Maximum size of the batch. The units are in number of
-            messages.
-        wait_time_seconds : int
-            Duration for which to wait. The units are in seconds.
-        message_factory : clare.common.messaging.factories.Message
-        """
-
-        self._sqs_queue = sqs_queue
-        self._batch_size_maximum_count = batch_size_maximum_count
-        self._wait_time_seconds = wait_time_seconds
-        self._message_factory = message_factory
-
-        self._buffer = _buffer if _buffer is not None else collections.deque()
-
-    def receive(self):
-        if not self._buffer:
-            self._fill_buffer()
-        try:
-            message = self._buffer.popleft()
-        except IndexError:
-            message = 'The receive operation timed out.'
-            raise consumer.exceptions.ReceiveTimeout(message)
-        return message
-
-    def _fill_buffer(self):
-        messages = self._sqs_queue.receive_messages(
-            MaxNumberOfMessages=self._batch_size_maximum_count,
-            WaitTimeSeconds=self._wait_time_seconds)
-
-        for message in messages:
-            marshalled = self._message_factory.create()
-
-            marshalled.id = message.message_id
-            marshalled.body = message.body
-            marshalled.delivery_receipt = message.receipt_handle
-
-            self._buffer.append(marshalled)
-
-    def __repr__(self):
-        repr_ = ('{}('
-                 'sqs_queue={}, '
-                 'batch_size_maximum_count={}, '
-                 'wait_time_seconds={}, '
-                 'message_factory={})')
-        return repr_.format(self.__class__.__name__,
-                            self._sqs_queue,
-                            self._batch_size_maximum_count,
-                            self._wait_time_seconds,
-                            self._message_factory)
