+++ downloadbot/infrastructure/consuming/tests/test_receivers.py
@@ -5,11 +5,40 @@ import itertools
 import queue
 
 import mock
-from nose.tools import assert_equal, assert_raises, assert_true
+from nose.tools import assert_equal, assert_raises, assert_true, raises
 
 from .. import receivers
 from downloadbot.common import messaging
 from downloadbot.common.messaging import consuming
+
+
+class MockSqsMessage:
+
+    def __init__(self, message_id, body, receipt_handle):
+        self.message_id = message_id
+        self.body = body
+        self.receipt_handle = receipt_handle
+
+    @classmethod
+    def from_message(cls, message):
+
+        """
+        Message adapter.
+
+        Parameters
+        ----------
+        message : downloadbot.common.messaging.messages.Message
+        """
+
+        return cls(message_id=message.id,
+                   body=message.body,
+                   receipt_handle=message.delivery_receipt)
+
+
+class MockSqsQueue:
+
+    def receive_messages(self):
+        pass
 
 
 class TestReceiver:
@@ -122,3 +151,34 @@ class TestConcurrentLinkedQueue(TestReceiver):
             side_effect=has_time_remaining)
         mock_countdown_timer.reset = mock.Mock()
         return mock_countdown_timer
+
+
+class TestSqsFifo(TestReceiver):
+
+    def test_receive_does_fill_if_buffer_is_empty(self):
+        return_value = [MockSqsMessage.from_message(self.message)]
+        sqs_queue = MockSqsQueue()
+        sqs_queue.receive_messages = mock.Mock(return_value=return_value)
+        receiver = receivers.SqsFifoQueue(sqs_queue=sqs_queue,
+                                          batch_size_maximum_count=None,
+                                          wait_time_seconds=None)
+        message = receiver.receive()
+        assert_equal(self.message.body, message.body)
+
+    def test_receive_does_not_fill_if_buffer_has_messages(self):
+        self._buffer.append(self.message)
+        receiver = receivers.SqsFifoQueue(sqs_queue=None,
+                                          batch_size_maximum_count=None,
+                                          wait_time_seconds=None,
+                                          _buffer=self._buffer)
+        message = receiver.receive()
+        assert_equal(self.message.body, message.body)
+
+    @raises(consuming.exceptions.ReceiveTimeout)
+    def test_receive_timeout_raises_exception(self):
+        sqs_queue = MockSqsQueue()
+        sqs_queue.receive_messages = mock.Mock(return_value=list())
+        receiver = receivers.SqsFifoQueue(sqs_queue=sqs_queue,
+                                          batch_size_maximum_count=None,
+                                          wait_time_seconds=None)
+        receiver.receive()
